Rename everything to avoid name clash with Vju GStreamerControl component
