+++ samples/GStreamerWPF/GStreamerD3D11.Demo/App.xaml
