+++ samples/GStreamerWPF/GStreamerD3D11.Demo/AssemblyInfo.cs
