+++ samples/GStreamerWPF/GStreamerD3D11.Demo/MainWindow.xaml.cs
