+++ samples/GStreamerWPF/GStreamerD3D11.Demo/Playback.cs
