+++ samples/GStreamerWPF/GStreamerD3D11.Demo/log4net.config
