+++ samples/GStreamerWPF/GStreamerD3D11/GStreamerCanvas.xaml
@@ -1,15 +1,15 @@
-﻿<UserControl x:Class="GStreamerControl.Library.GStreamerView"
+﻿<UserControl x:Class="GStreamerD3D11.GStreamerCanvas"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:GStreamerControl.Library"
+             xmlns:local="clr-namespace:GStreamerD3D11"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <UserControl.Resources>
         <BooleanToVisibilityConverter x:Key="booleanToVisibilityConverter"/>
     </UserControl.Resources>
     <Grid>
-        <Image x:Name="d3dScene" Visibility="{Binding Enabled, RelativeSource={RelativeSource Mode=FindAncestor, AncestorType=local:GStreamerView}, Converter={StaticResource booleanToVisibilityConverter}}"/>
+        <Image x:Name="d3dScene" Visibility="{Binding Enabled, RelativeSource={RelativeSource Mode=FindAncestor, AncestorType=local:GStreamerCanvas}, Converter={StaticResource booleanToVisibilityConverter}}"/>
     </Grid>
 </UserControl>
