+++ samples/GStreamerWPF/GStreamerD3D11/GStreamerD3D11.csproj
