+++ samples/GStreamerWPF/GStreamerD3D11/GStreamerD3D11.nuspec
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <package >
   <metadata>
-    <id>GStreamer.WPF</id>
+    <id>GStreamerD3D11</id>
     <version>1.1.0</version>
-    <title>GStreamer.WPF</title>
+    <title>GStreamerD3D11</title>
     <authors>berglie and iCsys AS</authors>
     <projectUrl>https://github.com/iCsysAS/gstreamer-d3d11-overlay</projectUrl>
     <description>Package for streaming composant in wpf</description>
@@ -16,6 +16,7 @@
   </metadata>
 	<files>
 		<file src="bin\Release\net6.0-windows\D3D11Scene.dll" target="lib\net6.0"></file>
+		<file src="bin\Release\net6.0-windows\GStreamerD3D11.dll" target="lib\net6.0"></file>
 		<file src="bin\Release\net6.0-windows\Interop.D3DImageEx.dll" target="lib\net6.0"></file>
 		<file src="bin\Release\net6.0-windows\Ijwhost.dll" target="lib\net6.0"></file>
 	</files>
